Moved index.js to the root directory
