+++ index.js
@@ -1,13 +1,13 @@
-import TheRoot from '../components/TheRoot.js'
-import { media }from '../store/media'
-import { lang }from '../store/lang'
-import { post }from '../store/post'
-import { page }from '../store/page'
-import { menu }from '../store/menu'
-import { meta }from '../store/meta'
-import { routes } from '../router/routes'
-import registerPlugin from './registerPlugin'
-import * as types from '../store/lang/mutation-types'
+import TheRoot from './components/TheRoot.js'
+import { media }from './store/media'
+import { lang }from './store/lang'
+import { post }from './store/post'
+import { page }from './store/page'
+import { menu }from './store/menu'
+import { meta }from './store/meta'
+import { routes } from './router/routes'
+import registerPlugin from './plugin/registerPlugin'
+import * as types from './store/lang/mutation-types'
 
 export const ModulePrefix = 'wp'
